requirements.txt
google-genai
matplotlib==3.7.2
numpy==1.24.3
python-dotenv==1.0.0
streamlit==1.32.0
pydantic==2.11.5